+++ gala/coordinates/orphan.py
@@ -3,14 +3,10 @@
 import astropy.coordinates as coord
 import astropy.units as u
 from astropy.coordinates import frame_transform_graph
-from astropy.coordinates.matrix_utilities import (rotation_matrix,
-                                                  matrix_product,
-                                                  matrix_transpose)
+from astropy.coordinates.matrix_utilities import rotation_matrix
 import numpy as np
 
-from gala.util import GalaDeprecationWarning
-
-__all__ = ["Orphan", "OrphanNewberg10", "OrphanKoposov19"]
+__all__ = ["OrphanNewberg10", "OrphanKoposov19"]
 
 
 class OrphanNewberg10(coord.BaseCoordinateFrame):
@@ -47,32 +43,37 @@ class OrphanNewberg10(coord.BaseCoordinateFrame):
         The Distance for this object along the line-of-sight.
 
     """
+
     default_representation = coord.SphericalRepresentation
     default_differential = coord.SphericalCosLatDifferential
 
     frame_specific_representation_info = {
         coord.SphericalRepresentation: [
-            coord.RepresentationMapping('lon', 'phi1'),
-            coord.RepresentationMapping('lat', 'phi2'),
-            coord.RepresentationMapping('distance', 'distance')]
+            coord.RepresentationMapping("lon", "phi1"),
+            coord.RepresentationMapping("lat", "phi2"),
+            coord.RepresentationMapping("distance", "distance"),
+        ]
     }
 
-    _default_wrap_angle = 180*u.deg
+    _default_wrap_angle = 180 * u.deg
 
     def __init__(self, *args, **kwargs):
-        wrap = kwargs.pop('wrap_longitude', True)
+        wrap = kwargs.pop("wrap_longitude", True)
         super().__init__(*args, **kwargs)
-        if wrap and isinstance(self._data, (coord.UnitSphericalRepresentation,
-                                            coord.SphericalRepresentation)):
+        if wrap and isinstance(
+            self._data,
+            (coord.UnitSphericalRepresentation, coord.SphericalRepresentation),
+        ):
             self._data.lon.wrap_angle = self._default_wrap_angle
 
     # TODO: remove this. This is a hack required as of astropy v3.1 in order
     # to have the longitude components wrap at the desired angle
-    def represent_as(self, base, s='base', in_frame_units=False):
+    def represent_as(self, base, s="base", in_frame_units=False):
         r = super().represent_as(base, s=s, in_frame_units=in_frame_units)
         if hasattr(r, "lon"):
             r.lon.wrap_angle = self._default_wrap_angle
         return r
+
     represent_as.__doc__ = coord.BaseCoordinateFrame.represent_as.__doc__
 
 
@@ -85,29 +86,32 @@ psi = 90.70 * u.degree
 D = rotation_matrix(phi, "z")
 C = rotation_matrix(theta, "x")
 B = rotation_matrix(psi, "z")
-R = matrix_product(B, C, D)
+R = B @ C @ D
 
 
-@frame_transform_graph.transform(coord.StaticMatrixTransform, coord.Galactic,
-                                 OrphanNewberg10)
+@frame_transform_graph.transform(
+    coord.StaticMatrixTransform, coord.Galactic, OrphanNewberg10
+)
 def galactic_to_orp():
-    """ Compute the transformation from Galactic spherical to
-        heliocentric Orphan coordinates.
+    """Compute the transformation from Galactic spherical to
+    heliocentric Orphan coordinates.
     """
     return R
 
 
 # Oph to Galactic coordinates
-@frame_transform_graph.transform(coord.StaticMatrixTransform, OrphanNewberg10,
-                                 coord.Galactic)
-def oph_to_galactic():
-    """ Compute the transformation from heliocentric Orphan coordinates to
-        spherical Galactic.
+@frame_transform_graph.transform(
+    coord.StaticMatrixTransform, OrphanNewberg10, coord.Galactic
+)
+def orp_to_galactic():
+    """Compute the transformation from heliocentric Orphan coordinates to
+    spherical Galactic.
     """
-    return matrix_transpose(galactic_to_orp())
+    return galactic_to_orp().T
 
 
 # ------------------------------------------------------------------------------
+
 
 class OrphanKoposov19(coord.BaseCoordinateFrame):
     """A coordinate frame for the Orphan stream defined by Sergey Koposov.
@@ -134,64 +138,57 @@ class OrphanKoposov19(coord.BaseCoordinateFrame):
 
     frame_specific_representation_info = {
         coord.SphericalRepresentation: [
-            coord.RepresentationMapping('lon', 'phi1'),
-            coord.RepresentationMapping('lat', 'phi2'),
-            coord.RepresentationMapping('distance', 'distance')]
+            coord.RepresentationMapping("lon", "phi1"),
+            coord.RepresentationMapping("lat", "phi2"),
+            coord.RepresentationMapping("distance", "distance"),
+        ]
     }
 
-    _default_wrap_angle = 180*u.deg
+    _default_wrap_angle = 180 * u.deg
 
     def __init__(self, *args, **kwargs):
-        wrap = kwargs.pop('wrap_longitude', True)
+        wrap = kwargs.pop("wrap_longitude", True)
         super().__init__(*args, **kwargs)
-        if wrap and isinstance(self._data, (coord.UnitSphericalRepresentation,
-                                            coord.SphericalRepresentation)):
+        if wrap and isinstance(
+            self._data,
+            (coord.UnitSphericalRepresentation, coord.SphericalRepresentation),
+        ):
             self._data.lon.wrap_angle = self._default_wrap_angle
 
     # TODO: remove this. This is a hack required as of astropy v3.1 in order
     # to have the longitude components wrap at the desired angle
-    def represent_as(self, base, s='base', in_frame_units=False):
+    def represent_as(self, base, s="base", in_frame_units=False):
         r = super().represent_as(base, s=s, in_frame_units=in_frame_units)
         if hasattr(r, "lon"):
             r.lon.wrap_angle = self._default_wrap_angle
         return r
+
     represent_as.__doc__ = coord.BaseCoordinateFrame.represent_as.__doc__
 
 
-@frame_transform_graph.transform(coord.StaticMatrixTransform,
-                                 coord.ICRS, OrphanKoposov19)
+@frame_transform_graph.transform(
+    coord.StaticMatrixTransform, coord.ICRS, OrphanKoposov19
+)
 def icrs_to_orp19():
-    """ Compute the transformation from ICRS to
-        heliocentric Orphan coordinates.
+    """Compute the transformation from ICRS to
+    heliocentric Orphan coordinates.
     """
-    R = np.array([[-0.44761231, -0.08785756, -0.88990128],
-                  [-0.84246097, 0.37511331, 0.38671632],
-                  [0.29983786, 0.92280606, -0.2419219]])
+    R = np.array(
+        [
+            [-0.44761231, -0.08785756, -0.88990128],
+            [-0.84246097, 0.37511331, 0.38671632],
+            [0.29983786, 0.92280606, -0.2419219],
+        ]
+    )
     return R
 
 
 # Oph to Galactic coordinates
-@frame_transform_graph.transform(coord.StaticMatrixTransform,
-                                 OrphanKoposov19, coord.ICRS)
-def oph19_to_icrs():
-    """ Compute the transformation from heliocentric Orphan coordinates to
-        spherical ICRS.
+@frame_transform_graph.transform(
+    coord.StaticMatrixTransform, OrphanKoposov19, coord.ICRS
+)
+def orp19_to_icrs():
+    """Compute the transformation from heliocentric Orphan coordinates to
+    spherical ICRS.
     """
-    return matrix_transpose(icrs_to_orp19())
-
-
-# TODO: remove this in next version
-class Orphan(OrphanNewberg10):
-    def __init__(self, *args, **kwargs):
-        import warnings
-        warnings.warn("This frame is deprecated. Use OrphanNewberg10 or "
-                      "OrphanKoposov19 instead.", GalaDeprecationWarning)
-        super().__init__(*args, **kwargs)
-
-
-trans = frame_transform_graph.get_transform(OrphanNewberg10,
-                                            coord.Galactic).transforms[0]
-frame_transform_graph.add_transform(Orphan, coord.Galactic, trans)
-trans = frame_transform_graph.get_transform(coord.Galactic,
-                                            OrphanNewberg10).transforms[0]
-frame_transform_graph.add_transform(coord.Galactic, Orphan, trans)
+    return icrs_to_orp19().T