+++ gala/coordinates/reflex.py
@@ -1,6 +1,6 @@
 import astropy.coordinates as coord
 
-__all__ = ['reflex_correct']
+__all__ = ["reflex_correct"]
 
 
 def reflex_correct(coords, galactocentric_frame=None):
@@ -38,6 +38,6 @@ def reflex_correct(coords, galactocentric_frame=None):
 
     observed = c.transform_to(galactocentric_frame)
     rep = observed.cartesian.without_differentials()
-    rep = rep.with_differentials(observed.cartesian.differentials['s'] + v_sun)
+    rep = rep.with_differentials(observed.cartesian.differentials["s"] + v_sun)
     fr = galactocentric_frame.realize_frame(rep).transform_to(c.frame)
     return coord.SkyCoord(fr)